refactor: extract shared search module, unify embedding client (#59)
Extract cosine_similarity, embed_query, embed_texts, and
synthesize_project_context into scripts/lib/search_core.py. The module
delegates embedding to gemini_embeddings.py (the single Gemini client)
rather than reimplementing the HTTP call.

For standalone use via /focus, search_core.py and gemini_embeddings.py
are vendored into skills/focus/scripts/. A drift check script ensures
vendored copies stay in sync with canonical sources.

Key decisions:
- No sys.exit in library code — errors propagate as exceptions
- One embedding client (gemini_embeddings.py), not two
- Thin adapter in search_core, not a reimplementation
- Net -136 LOC across the two search scripts

Closes #59
